requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
coloredlogs==15.0.1
e==1.4.5
humanfriendly==10.0
idna==3.3
mongoengine==0.23.1
multidict==5.2.0
psutil==5.8.0
git+git://github.com/Pycord-Development/pycord@1e77d6e#egg=py-cord
pymongo==3.12.1
pytimeparse==1.1.8
python-dotenv==0.19.1
typing-extensions==3.10.0.2
yarl==1.7.0
